got FF to work but QUOTE seems to be a problem
look at theoretical interpreters section to see an explenation of sorts:
https://web.archive.org/web/20041209025920/http://www8.informatik.uni-erlangen.de/html/lisp/mcc91.html

this might be important:
"This changed the substitution procedure into a frozen one -- the a-list was used as in the running system. The interpreter works with a mixture of evaluated and unevaluated arguments. Top level is {\tt apply}. The arguments are quoted and {\tt eval} is activated. Argument subterms are not evaluated if the function is written as lambda-expression. Basic functions have to evaluate their arguments. If there is an expression (or a variable as function name) in functional position, arguments are evaluated. After argument evaluation, the values are quoted. This way, arguments do not get evaluated twice -- but {\tt evlis} strips quotes off and puts them on again: a waste of storage. In the a-list, quoted values and unevaluated arguments are mixed. On access, the values get evaluated in the actual environment, which might be quite different than the original one.""